get AcquireError to public
