+++ errors.go
@@ -2,6 +2,8 @@ package ydb
 
 import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/errors"
+	ratelimiterErrors "github.com/ydb-platform/ydb-go-sdk/v3/internal/ratelimiter/errors"
+	"github.com/ydb-platform/ydb-go-sdk/v3/ratelimiter"
 )
 
 func IterateByIssues(err error, it func(message string, code uint32, severity uint32)) {
@@ -113,4 +115,12 @@ func IsOperationErrorNotFoundError(err error) bool {
 
 func IsOperationErrorSchemeError(err error) bool {
 	return IsOperationErrorCode(err, int32(errors.StatusSchemeError))
+}
+
+func IsRatelimiterAcquireError(err error) bool {
+	return ratelimiterErrors.IsAcquireError(err)
+}
+
+func ToRatelimiterAcquireError(err error) ratelimiter.AcquireError {
+	return ratelimiterErrors.ToAcquireError(err)
 }
