+++ ratelimiter/errors.go
@@ -1,23 +1,8 @@
 package ratelimiter
 
-import "fmt"
+type AcquireError interface {
+	error
 
-type acquireError struct {
-	err    error
-	amount uint64
-}
-
-func (e *acquireError) Error() string {
-	return fmt.Sprintf("acquire for amount (%d) failed: %v", e.amount, e.err)
-}
-
-func (e *acquireError) Unwrap() error {
-	return e.err
-}
-
-func AcquireError(amount uint64, err error) error {
-	return &acquireError{
-		err:    err,
-		amount: amount,
-	}
+	Amount() uint64
+	Unwrap() error
 }
